requirements.txt
pandas
matplotlib
seaborn
wordcloud
pyecharts
spacy
biopython
numpy

# Optional (install if needed):
# jupyter
# openpyxl